+++ frontend/omni/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - User avatar at the bottom of the sidebar showing the logged-in user.
+- Logout button in the sidebar footer.
 
 ## [0.0.2] - 2026-04-13